fix(condo): DOMA-11237 move custom styles to global css
