--- apps/condo/domains/property/components/panels/Builder/BuildingEditTopMenu.css
@@ -1,7 +1,0 @@
-.add-element-dropdown > .condo-dropdown-menu {
-  max-height: 100vh;
-}
-
-.add-element-dropdown > .condo-dropdown-menu .condo-dropdown-menu-title-content > .condo-space {
-  align-items: center;
-}
