--- apps/condo/domains/property/components/panels/Builder/BuildingPanelEdit.css
@@ -1,3 +1,0 @@
-.map-edit-address-container .condo-select > .condo-select-selector {
-  background-color: transparent;
-}
