--- apps/condo/domains/property/components/panels/Builder/BuildingPanelView.css
@@ -1,35 +1,0 @@
-.property-map-container {
-  width: 100%;
-  height: 100%;
-  text-align: center;
-}
-
-.property-map-scroll-holder {
-  position: static;
-  padding-bottom: 10px;
-  white-space: nowrap;
-}
-
-.property-map-scroll-container {
-  width: 100%;
-  padding-bottom: 20px;
-  overflow-y: hidden;
-}
-
-.unit-button-tooltip-wrapper {
-  display: inline-flex;
-  margin-right: 8px;
-}
-
-.property-map-floor-container {
-  display: block;
-}
-
-.property-map-unit-type-row {
-  padding-left: 8px;
-}
-
-.property-map-fullscreen-header {
-  align-items: center;
-  margin-bottom: 28px;
-}
