--- apps/condo/domains/property/components/panels/Builder/UnitButton/UnitButton.css
@@ -1,99 +1,0 @@
-.map-unit-button {
-  width: 44px;
-  min-width: 44px;
-  height: 44px;
-  margin-top: 8px;
-  margin-right: 8px;
-  padding: 0 8px;
-  color: var(--condo-global-color-black);
-  background-color: var(--condo-global-color-white);
-  border: 1px solid var(--condo-global-color-gray-1);
-  transition: var(--condo-global-transition-duration-default);
-}
-
-.map-unit-button:last-child {
-  margin-right: 0;
-}
-
-.map-unit-button:focus,
-.map-unit-button:hover {
-  color: var(--condo-global-color-black);
-  border-color: var(--condo-global-color-gray-7);
-}
-
-.map-unit-button:disabled,
-.map-unit-button:disabled:hover {
-  color: black;
-  background-color: var(--condo-global-color-white);
-  border: transparent;
-  cursor: default;
-  pointer-events: none;
-}
-
-.condo-btn.map-unit-button-section {
-  background-color: transparent;
-  border-color: var(--condo-global-color-gray-5);
-}
-
-.map-unit-button-section:hover,
-.map-unit-button-section:focus {
-  background-color: transparent;
-  border-color: var(--condo-global-color-gray-7);
-}
-
-.map-unit-button.map-unit-button-floor {
-  display: block;
-  background-color: transparent;
-  border: none;
-}
-
-.map-unit-button-section > .condo-btn-text,
-.map-unit-button-floor > .condo-btn-text {
-  font-size: 14px;
-}
-
-.condo-btn.map-unit-button.map-unit-button-selected {
-  background: var(--condo-global-color-brand-gradient-1);
-  border-color: var(--condo-global-color-green-5);
-}
-
-.map-unit-button > .condo-btn-text:hover,
-.map-unit-button > .condo-btn-text:focus {
-  background: none;
-}
-
-.condo-btn.map-unit-button.map-unit-button-selected > .condo-btn-text,
-.condo-btn.map-unit-button.map-unit-button-selected:hover > .condo-btn-text,
-.condo-btn.map-unit-button.map-unit-button-selected:focus > .condo-btn-text {
-  color: var(--condo-global-color-green-7);
-}
-
-.map-preview-unit-button {
-  opacity: 0.5;
-}
-
-.condo-btn.map-unit-button.map-duplicate-unit-button {
-  background-color: var(--condo-global-color-red-5);
-}
-
-.map-unit-button.map-flat-unit-button,
-.map-unit-button.map-flat-unit-button:hover,
-.map-unit-button.map-parking-unit-button,
-.map-unit-button.map-parking-unit-button:hover {
-  background-color: var(--condo-global-color-white);
-}
-
-.map-unit-button.map-apartment-unit-button,
-.map-unit-button.map-apartment-unit-button:hover {
-  background-color: #cbe2f6;
-}
-
-.map-unit-button.map-warehouse-unit-button,
-.map-unit-button.map-warehouse-unit-button:hover {
-  background-color: #f086334d;
-}
-
-.map-unit-button.map-commercial-unit-button,
-.map-unit-button.map-commercial-unit-button:hover {
-  background-color: #eb34684d;
-}
